pyproject.toml
[project]
name = "nexus-pcv"
version = "0.2.1"
description = "A CLI tool to perform a pre-change validation on Nexus Dashboard Insights or Network Assurance Engine."
license = { text = "MPL-2.0" }
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Daniel Schmidt", email = "danischm@cisco.com" },
]
maintainers = [
    { name = "Daniel Schmidt", email = "danischm@cisco.com" },
]
dependencies = [
    "typer>=0.15",
    "pyyaml>=6.0.2",
    "httpx>=0.28",
]

[project.urls]
homepage = "https://github.com/netascode/nexus-pcv"
repository = "https://github.com/netascode/nexus-pcv"
documentation = "https://github.com/netascode/nexus-pcv"

[project.scripts]
nexus-pcv = "nexus_pcv.cli.main:app"

[project.optional-dependencies]
dev = [
    "bandit[toml]>=1.7.0",
    "mypy>=1.17",
    "pre-commit>=4.0",
    "pytest>=8.4",
    "pytest-cov>=6.0",
    "pytest-mock>=3.15",
    "ruff>=0.12",
    "types-pyyaml>=6.0.12.20250822",
]

[tool.coverage.run]
source = ["nexus_pcv"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "@abstract", "except ImportError:"]
omit = ["*/__main__.py"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
python_version = "3.13"
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = false

[tool.pytest.ini_options]
markers = ["unit", "integration"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings  
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["nexus_pcv"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert_used test

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
